Reorganize files to get rid of pointless folders
