+++ client/src/components/ModuleCreator.tsx
@@ -1,8 +1,8 @@
 import React, { useState } from "react";
 import Modal from "@mui/material/Modal";
-import { useUser } from "../../hooks/useUser";
-import { ApiClient } from "../../hooks/ApiClient";
-import { ModuleCreatorProps } from "../../types";
+import { useUser } from "../hooks/useUser";
+import { ApiClient } from "../hooks/ApiClient";
+import { ModuleCreatorProps } from "../types";
 
 function ModuleCreator({ addModule }: ModuleCreatorProps) {
   const [moduleName, setModuleName] = useState("");
