+++ client/src/components/Navbar.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Link, useNavigate } from "react-router-dom";
-import { useUser } from "../../hooks/useUser";
-import { ApiClient } from "../../hooks/ApiClient";
+import { useUser } from "../hooks/useUser";
+import { ApiClient } from "../hooks/ApiClient";
 
 function NavBar() {
   return (
