+++ client/src/screens/DefaultContainer.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import NavBar from '../../components/NavBar/Navbar';
+import NavBar from '../components/Navbar';
 import { Outlet } from 'react-router';
 
 function DefaultScreen() {
