--- client/src/screens/LearningPlan/LearningPlan.css
@@ -1,4 +1,0 @@
-div { 
-    padding: "2%";
-    height: "100vh"; 
-}
